[SMT-COMP] Workaround for `tear-down-incremental`
This commit implements a temporary fix for the
`--tear-down-incremental=X` option. The issue is that `reset()` now
replaces the `Solver` instance being used. However, the symbol table of
the parser holds references to `Expr`s from the old `ExprManager` (and
the pointer to `SmtEngine` in `CommandExecutor` is stale after
replacing the `Solver` but that is an easy fix). For the competition,
the easiest fix is to call `SmtEngine::reset()`. In experiments,
`--tear-down-incremental=X` performs significantly better and avoids
correctness issues.