SAT-ctx dependent proofs of assertions to the EE
Previously, proofs that ArithCongruenceMaganer exposed to the equality
engine were stored in a user-context-dependent EagerPfGen.

Now, that generator is SAT-context dependent.

This is important, because the proof are open: in terms of literals that
are valid in the current SAT context. During SAT trail pops, proofs may
need to be discarded.