Add testing infrastructure for LFSC signatures
This commit adds testing infrastructure for LFSC signatures that is
enabled when CVC4 is configured with LFSC. The testing infrastructure
adopts `run_test.py` from https://github.com/CVC4/LFSC with minor
modifications (mainly adding support for a list of include directories
that are searched to resolve *.plf dependencies). The commit uses the
existing examples and test files from `proofs/signatures` as the initial
set of tests.

Co-authored-by: Alex Ozdemir <aozdemir@hmc.edu>